Rewrite the JS so it gives a fuck.
The JS was kind of terrible. Assigning based on increasing a literal index
doesn't make sense, and the if 0, set 1 thing was absolutely not neccessary
since we were assigning a nofucks image to the 0 index of the images array.
By just getting a single damned random number based upon the length of the
images array and looking it up by that is really sufficient enough for the
incomplexity of getting a random image from a static array.

Also, make it an anonymous immediately invoked function expression, since we
aren't calling this thing multiple times, and the whole reload thing just
reloads the page, not rerun the JS inline.